+++ lib/src/event/event.dart
@@ -331,6 +331,19 @@ abstract class Event<H> {
   }
   
   /**
+   * Sets the {@link EventListener} to receive events for the given element.
+   * Only one such listener may exist for a single element.
+   * 
+   * @param elem the element whose listener is to be set
+   * @param listener the listener to receive {@link Event events}
+   */
+  static void setEventListener(dart_html.Element elem, EventListener listener) {
+    // This cast is always valid because both Element types are JSOs and have
+    // no new fields are added in the subclass.
+    Dom.setEventListener(elem, listener);
+  }
+  
+  /**
    * Sets the current set of events sunk by a given element. These events will
    * be fired to the nearest {@link EventListener} specified on any of the
    * element's parents.