+++ lib/util.dart
@@ -19,3 +19,4 @@ part 'src/util/safe_html.dart';
 part 'src/util/safe_html_utils.dart';
 part 'src/util/safe_html_string.dart';
 part 'src/util/layout_constant.dart';
+part 'src/util/animation_type.dart';